+++ packages/network-connection-banner-controller/CHANGELOG.md
@@ -10,5 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial release ([#9041](https://github.com/MetaMask/core/pull/9041))
+- Add an explicit, idempotent `init()` lifecycle method that starts banner
+  evaluation after dependent controllers are ready
+  ([#9041](https://github.com/MetaMask/core/pull/9041))
 
 [Unreleased]: https://github.com/MetaMask/core/